Linux alpha test
git-svn-id: https://svn.pjsip.org/repos/pjproject/main@35 74dad513-b988-da41-8d7b-12977e46ad98

+++ pjlib/src/pj/os_core_unix.c
@@ -260,7 +260,7 @@ static void *thread_main(void *param)
     PJ_LOG(6,(rec->obj_name, "Thread started"));
 
     /* Call user's entry! */
-    result = (void*) (*rec->proc)(rec->arg);
+    result = (void*)(long)(*rec->proc)(rec->arg);
 
     /* Done. */
     PJ_LOG(6,(rec->obj_name, "Thread quitting"));
